feat: adicionado a funcionalidade de editar e deletar

+++ src/components/deletarTransacao/deletar.tsx
@@ -8,27 +8,33 @@ import {
   DialogHeader,
   DialogTitle,
 } from "../ui/dialog";
+import type { Transacao } from "@/app/type/type";
 
 interface propsDelete {
   open: boolean;
-  onOpenChange: (value: boolean) => void;
+  onOpenChange: (open: boolean) => void;
+  transacao: Transacao | null;
 }
 
-export const DeletarTransacao = ({ open, onOpenChange }: propsDelete) => {
-  const deletarTransacao = async (id: string, onSucess: () => void) => {
+export const DeletarTransacao = ({
+  open,
+  onOpenChange,
+  transacao,
+}: propsDelete) => {
+  const deletar = async (id: number | undefined) => {
+    if (id === undefined) return;
     try {
       await api.delete(`/transaction/${id}`);
-      alert("Deletado com sucesso");
-      onSucess();
+      onOpenChange(false);
+      window.location.reload();
     } catch {
       alert("Erro ao deletar");
     }
   };
+
   return (
-    <div>
-      <Trash2 className=" h-4 w-4" />
-      <Dialog open={open} onOpenChange={onOpenChange}>
-     <DialogContent className="w-[90vw] max-w-sm rounded-xl border-[0.5px] border-[#ebebeb] bg-white p-0 shadow-[0_8px_40px_rgba(0,0,0,0.10)]">
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent className="w-[90vw] max-w-sm rounded-xl border-[0.5px] border-[#ebebeb] bg-white p-0 shadow-[0_8px_40px_rgba(0,0,0,0.10)]">
         <DialogHeader className="border-b border-[#f5f5f3] px-4 pb-1 pt-2 items-center">
           <div className="flex items-center justify-center m-5">
             <TriangleAlert className="h-6 w-6 text-[#9a9a94]" />
@@ -49,7 +55,7 @@ export const DeletarTransacao = ({ open, onOpenChange }: propsDelete) => {
             Cancelar
           </button>
           <button
-            
+            onClick={() => deletar(transacao?.id)}
             className="flex-1 border border-red-500 bg-red-500 text-white py-2 px-4 rounded-lg hover:bg-red-600"
           >
             Deletar
@@ -57,6 +63,5 @@ export const DeletarTransacao = ({ open, onOpenChange }: propsDelete) => {
         </DialogFooter>
       </DialogContent>
     </Dialog>
-    </div>
   );
 };
